Implement WebSocket log streaming endpoint and handlers
The changes add WebSocket support for real-time log streaming with proper
connection handling, message routing, and server-side echo functionality.
