+++ src/pages/api/load_tests/mod.rs
@@ -11,4 +11,5 @@ pub fn load_tests_scope() -> Scope {
         .route("", web::post().to(post_new_loadtest::new_loadtest))
         .route("/new", web::get().to(get_load_test_from))
         .route("/{id}/logs", web::get().to(get_logs_stream::logs_stream))
+        .route("/{id}/logs/ws", web::get().to(get_log_stream_ws::log_stream_ws))
 }
