+++ src/services/websocket/mod.rs
@@ -5,3 +5,4 @@ pub type Msg = String;
 
 pub mod handler;
 pub mod server;
+pub mod server_handler;
