+++ src/services/websocket/server.rs
@@ -3,10 +3,10 @@ use rand::Rng;
 use std::collections::{HashMap, HashSet};
 use tokio::sync::{mpsc, oneshot};
 
-use super::{ConnId, Msg, RoomId};
+use super::{ConnId, Msg, RoomId, server_handler::LogServerHandler};
 
 #[derive(Debug)]
-enum Command {
+pub enum Command {
     Connect {
         conn_tx: mpsc::UnboundedSender<Msg>,
         res_tx: oneshot::Sender<ConnId>,
@@ -94,38 +94,5 @@ impl LogServer {
             }
         }
         Ok(())
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct LogServerHandler {
-    cmd_tx: mpsc::UnboundedSender<Command>,
-}
-impl LogServerHandler {
-    pub async fn connect(&self, conn_tx: mpsc::UnboundedSender<Msg>, log_id: i32) -> ConnId {
-        let (res_tx, res_rx) = oneshot::channel();
-        self.cmd_tx
-            .send(Command::Connect {
-                conn_tx,
-                res_tx,
-                log_id,
-            })
-            .unwrap();
-
-        res_rx.await.unwrap()
-    }
-    pub fn disconnect(&self, conn: ConnId) {
-        self.cmd_tx.send(Command::Disconnect { conn }).unwrap();
-    }
-    pub async fn send_msesage(&self, log_id: i32, msg: impl Into<Msg>) {
-        let (res_tx, res_rx) = oneshot::channel();
-        self.cmd_tx
-            .send(Command::Message {
-                log_id,
-                msg: msg.into(),
-                res_tx,
-            })
-            .unwrap();
-        res_rx.await.unwrap();
     }
 }
